requirements/base.txt
boto3==1.7.13
botocore==1.10.13
coreapi==2.3.3
Django==2.0.9
django-easy-select2==1.4.0
django-fsm==2.6.0
django-fsm-admin==1.2.4
django-import-export==1.0.1
django-prettyjson==0.4.0
django-rest-knox==3.1.4
django-storages==1.6.6
django-webpack-loader==0.6.0
djangorestframework==3.8.2
factory-boy==2.10.0
Faker==0.8.13
mccabe==0.6.1
mysqlclient==1.3.12
pandas==0.23.0
Pillow==5.0.0
psycopg2-binary==2.7.4
python-decouple==3.1
pytz==2018.4
